+++ tranquility-types/Cargo.toml
@@ -6,10 +6,11 @@ edition = "2018"
 license = "MIT"
 
 [dependencies]
+chrono = { version = "0.4.19", optional = true }
 serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0.64"
 
 [features]
-activitypub = [ ]
+activitypub = ["chrono"]
 mastodon = [ ]
 webfinger = [ ]
